final changes before merging
This commit starts by containing the separation between the post and the get methods that will be used for handling crowdactions
